backend/requirements.txt
azure-functions>=1.17.0
azure-data-tables>=12.4.0